Refactored and removed old Point2D.
