+++ aima-core/src/main/java/aima/core/environment/map/Map.java
@@ -2,7 +2,7 @@ package aima.core.environment.map;
 
 import java.util.List;
 
-import aima.core.util.datastructure.Point2D;
+import aima.core.util.math.geom.shapes.Point2D;
 
 /**
  * Provides a general interface for maps.
