+++ aima-gui/src/main/java/aima/gui/swing/demo/agent/map/MapAgentView.java
@@ -10,7 +10,7 @@ import aima.core.agent.Agent;
 import aima.core.agent.Environment;
 import aima.core.environment.map.Map;
 import aima.core.environment.map.MapEnvironment;
-import aima.core.util.datastructure.Point2D;
+import aima.core.util.math.geom.shapes.Point2D;
 import aima.gui.swing.framework.EmptyEnvironmentView;
 
 /**
